suppression d'un Book ok

+++ BackEnd/Controllers/book.js
@@ -1,4 +1,5 @@
 const Book = require('../models/Book')
+const fs= require('fs');
 
 exports.createBook= (req, res, next) => {
   const bookObject=JSON.parse(req.body.book)
@@ -36,11 +37,24 @@ exports.createBook= (req, res, next) => {
         });
  };
 
-  exports.deleteBook= (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-      .catch(error => res.status(400).json({ error }));
-  }
+ exports.deleteBook = (req, res, next) => {
+  Book.findOne({ _id: req.params.id})
+      .then(book => {
+          if (book.userId != req.auth.userId) {
+              res.status(401).json({message: 'Not authorized'});
+          } else {
+              const filename = book.imageUrl.split('/images/')[1];
+              fs.unlink(`images/${filename}`, () => {
+                  Book.deleteOne({_id: req.params.id})
+                      .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                      .catch(error => res.status(401).json({ error }));
+              });
+          }
+      })
+      .catch( error => {
+          res.status(500).json({ error });
+      });
+};
 
   exports.getAllBooks= (req, res, next) => {
     Book.find()
